feat: Add keyboard shortcuts for moving selected todos up and down
The code changes introduce new keyboard shortcuts for moving the selected todo item up and down in the list. This will improve the user experience by providing an alternative way to navigate through the todos without changing the focus.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature implementation.