refactor: Sort todos by index in TodoController
This commit adds sorting functionality to the TodoController by sorting the todos list based on the index property of each AppTodoItem. The sorting is done in ascending order using the compareTo method. This change ensures that the todos are displayed in the correct order according to their index.

Note: This commit message follows the established convention of using a prefix (refactor) to indicate code refactoring.