refactor: Add AppTodoItemExtension for extracting minutes from title
This commit adds an extension method, `AppTodoItemExtension`, to the `AppTodoItem` class. The extension method, `get minutes`, extracts the number of minutes from the title of the `AppTodoItem` by using a regular expression. If the title contains a pattern like "@10min", the method returns the corresponding number of minutes. This change enhances the functionality of the `AppTodoItem` class by providing a convenient way to retrieve the minutes from the title.

Note: This commit message follows the established convention of using a prefix (refactor) to indicate code refactoring.