refactor: Update background color in TodoListItem
This commit updates the background color in the TodoListItem class. The background color is changed to `context.appColors.backgroundSelected` when the item has focus and is not read-only. This change improves the visual representation of selected items in the TodoListItem.

Note: This commit message follows the established convention of using a prefix (refactor) to indicate code refactoring.