feat: Add google_client_id and google_reversed_client_id to dev.json
This commit adds the `google_client_id` and `google_reversed_client_id` fields to the `dev.json` file in the `desktop/dart_defines` package. These fields are used for Google authentication. By including these values in the configuration file, the app will have the necessary credentials to authenticate with Google.

[Reference issue #123]