chore: Update Firebase dependencies to version 10.25.0
This commit updates the Firebase dependencies in the Podfile and Podfile.lock files to version 10.25.0. The FirebaseCoreExtension dependency is removed, and the FirebaseFirestore dependency is updated to use the GitHub repository with the specified tag. This change ensures that the app uses the latest version of Firebase libraries for improved functionality and compatibility.

[Reference issue #456]